Separating out search into files.
