+++ public/src/js/search/collections/OppsCollections.js
@@ -1,0 +1,3 @@
+module.exports = Backbone.Collection.extend({
+  models: SearchShop.Model.Item
+});
