+++ public/src/js/search/models/OppModel.js
@@ -1,0 +1,10 @@
+module.exports = Backbone.Model.extend({
+  defaults: {
+    title: '',
+  },
+  // Increase or decrease the quantity
+	quanity : function( type ) {
+		var qty = this.get('quantity');
+		this.set('quantity', (type === 'increase' ? ++qty : --qty) );
+	}
+});
